Add files via upload
This python script is designed to show the hierarchical relationships of affected phenotypic systems. After running it with python, it will give you a window. By clicking on desired MP ID and computing the ancestor tree of this MP, it would give you the result. 
This code reveals something good because it makes visible clearly the hairball dataset. Visualization is an important issue for being able to see the relationships, so this code is good in this way.